fixed bug in simple_read and simple_write when the number of elements and/or the number of iteration was too big to be contained in an int

+++ unit_tests/simple_read.cpp
@@ -16,15 +16,15 @@ int sum_all(int *data, int num_elements, int num_reads) {
 
 int main (int argc, char** argv) {
 	int rank;
-	int num_elements, num_reads;
+	long int num_elements, num_reads;
 	MPI_Init(&argc, &argv);
 	if (argc != 3) {
 		std::cout << "simple read input error" << std::endl;
 		MPI_Finalize();
 		return 0;
 	}
-	num_elements = std::atoi(argv[1]);
-	num_reads = std::atoi(argv[2]);
+	num_elements = std::atol(argv[1]);
+	num_reads = std::atol(argv[2]);
 	MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 	std::string file_name = "file_" + std::to_string(rank) + ".txt";
 	int fd = open(file_name.c_str(), O_RDONLY, 0644);
